Reorder classes into packages, move methods around, introduce Relationships
This update will bring some order to the chaos that is the orderchaos package. Methods are now (mostly) neatly ordered in packages, and some methods have shifted from Being to Simulation as a first step to crawl back to a neat MVC pattern.
Besides this, a Relationship class has been added to replace the Partner in Being, making it more accessible from the Simulation. This has caused some changes in a number of methods lying around, such as the MakeChild method.