Add Apocalypse, revamp matchmaking system, split population
This update introduces the apocalypse. This will give you the ability to just kill everyone (or spare two people) and start over.
Besides this, the matchmaking system has had an overhaul which means siblings are no longer able to fall in love with eachother.
Lastly, the population set has been split up into two sets: a living one and a dead one, for reasons.